validation file exist, before delete

+++ src/controllers/highlineController.js
@@ -112,7 +112,9 @@ export const deleteImage = async (req, res) => {
         if (!images[0])
             throw new AppError("no image found.", 404);
 
-        await fs.unlink(`uploads/${path.basename(images[0])}`, resultHandler);
+        if (fs.exists(`uploads/${path.basename(images[0])}`)) {
+            await fs.unlink(`uploads/${path.basename(images[0])}`, resultHandler);
+        }
 
         images = higlineFound.imagesUrl.filter(urlValue => urlValue != imgUrl);
